requirements.txt
Kivy==2.3.0
numpy==1.26.4
cvzone== 1.6.1
opencv-python==4.9.0.80
google-generativeai==0.8.3